pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fm_shiny_app"
version = "0.1"
requires-python = ">= 3.12"
dependencies = [
    "rsconnect",
    "pandas",
    "pyjanitor",
    "pyprojroot",
    "shiny"
]

[tool.setuptools.packages.find]
where = ["src"]